+++ modules/admin/controllers/SlidertopController.php
@@ -113,9 +113,15 @@ class SlidertopController extends AdminController {
             $id = (int)Yii::$app->request->getQueryParams()['id'];
 
             $slides = Slides::findOne($id);
-            if ($slides->delete() !== false) {
-                $response = true;
+            if ($slides) {
+                $image = $slides->image;
+                if ($slides->delete() !== false) {
+                    unlink(Yii::$app->basePath . '/web' . Yii::$app->params['params']['pathToImage'] . Slides::IMG_FOLDER_SLIDER_TOP . 'admin_' . $image);
+                    unlink(Yii::$app->basePath . '/web' . Yii::$app->params['params']['pathToImage'] . Slides::IMG_FOLDER_SLIDER_TOP . $image);
+                    $response = true;
+                }
             }
+
 
             Yii::$app->response->format = Response::FORMAT_JSON;
             return [
